+++ src/models/VCardSettings.php
@@ -1,4 +1,12 @@
 <?php
+/**
+ * vCard plugin for Craft CMS 4.x
+ *
+ * vCard generator plugin for Craft CMS 4
+ *
+ * @link      http://n43.me
+ * @copyright Copyright (c) 2022 Nathaniel Hammond (nfourtythree)
+ */
 
 namespace nfourtythree\vcard\models;
 
@@ -7,5 +15,5 @@ use Yii;
 
 class VCardSettings extends Model
 {
-    public $salt = null;
+    public ?string $salt = null;
 }