+++ src/models/VCard_UrlModel.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * vCard plugin for Craft CMS 3.x
+ * vCard plugin for Craft CMS 4.x
  *
- * vCard generator plugin for Craft cms 3
+ * vCard generator plugin for Craft CMS 4
  *
  * @link      http://n43.me
- * @copyright Copyright (c) 2019 Nathaniel Hammond (nfourtythree)
+ * @copyright Copyright (c) 2022 Nathaniel Hammond (nfourtythree)
  */
 
 namespace nfourtythree\vcard\models;
@@ -18,40 +18,25 @@ use craft\base\Model;
 /**
  * VCard_UrlModel Model
  *
- * Models are containers for data. Just about every time information is passed
- * between services, controllers, and templates in Craft, it’s passed via a model.
- *
- * https://craftcms.com/docs/plugins/models
- *
  * @author    Nathaniel Hammond (nfourtythree)
  * @package   VCard
  * @since     1.0.0
  */
 class VCard_UrlModel extends Model
 {
-    // Public Properties
-    // =========================================================================
-    public $address = '';
-    public $type = '';
-
-    // Public Methods
-    // =========================================================================
+    public string $address = '';
+    public string $type = '';
 
     /**
-     * Returns the validation rules for attributes.
-     *
-     * Validation rules are used by [[validate()]] to check if attribute values are valid.
-     * Child classes may override this method to declare different validation rules.
-     *
-     * More info: http://www.yiiframework.com/doc-2.0/guide-input-validation.html
-     *
-     * @return array
+     * @inheritdoc
      */
-    public function rules()
+    public function defineRules(): array
     {
-        return [
-          [ 'address' ,'string' ],
-          [ 'type' ,'string' ],
-        ];
+        $rules = parent::defineRules();
+
+        $rules[] = ['address', 'string'];
+        $rules[] = ['type', 'string'];
+
+        return $rules;
     }
 }